Cleanup.

+++ asteria/src/runtime/variable.hpp
@@ -82,9 +82,9 @@ class Variable : public virtual RefCnt_Base
       {
         this->m_gcref_mant += dmant;
         // Add with carry.
-        const auto carry = static_cast<long>(this->m_gcref_mant);
+        const auto carry = static_cast<int>(this->m_gcref_mant);
         this->m_gcref_intg += carry;
-        this->m_gcref_mant -= static_cast<double>(carry);
+        this->m_gcref_mant -= carry;
       }
   };
